Reword javap filtering options
"To filter out" means remove, and, therefore, has the exact opposite meaning of what the example shows. But "filter" is kind of an ambiguous word when it comes to what's removed and what's left, not to mention which you are interested in (if you filter water to drink that's one thing, if you filter water to mine gold, that's another), so I'm avoiding it altogether and going for some unambiguous.