SI-6476 Promote the puzzling esc err to parser
Give the parser enough knowledge of `s` and `f`
to emit an error on the puzzling badly escaped
quote.  The error is emitted as supplemental
advice when a subsequent syntax error is encountered,
such as an unclosed literal.

The test is for a `\` at the end of a part, which
can be written explicitly (because the user doesn't
know how to escape `\\`), or inadvertently (because
of `s"abc\"def"` or `s"abc\$x"` where something else
is probably intended.