SI-8608 f interpolator emits plain old strings
When invoking `format` is obviated by a lack of
formatting fields, then just degenerate to an
unenhanced string.

This means it doesn't cost anything to use
f"$$ordinary" in place of "$ordinary", which
may cause warnings under -Xlint.

Note that certain format literals, in particular
for line separator %n, are not actually literals and
can't be replaced at compile time.