Route type checker reporting through context
Continue the work started in SI-8450
(no "implicit numeric widening" warning in silent mode),
which was caused by going straight to the reporter instead of
using the context for type error reporting (which buffers in silent mode).

Ideally, this mistake should not be possible: typer should change
the current reporter to buffer where appropriate.