Regularize `comment` hook method
This is currently pretty borken,
but let's at least not clutter innocent
interfaces with this functionality.

Moved `comment` (as `signalParsedDocComment`) next to
the other hook methods in `Global`. For now, it calls
the old `reporter.comment` hook method.

As soon as the IDE is refactored to receive comments properly,
the deprecated `Reporter#comment` method can be removed.