Concretize diagnostics: one boolean suffices for now.
Once we get the next diagnostic, lets encapsulate them in an
object, with a boolean flag for each one when it needs to trigger,
and a nice message that should be presented to our delighted user.

A list of Strings that is searched by contents is a bit fragile,
and can't be very fast either.