Untangle reporting of ambiguous errors.
Now that all reporting mode manipulators are private to Context,
let's untangle this logic:
  - every `setReportErrors` gets a corresponding `setAmbiguousErrors(true)`
  - every `setBufferErrors` gets a corresponding `setAmbiguousErrors(false)`
  - every `setThrowErrors`  gets a corresponding `setAmbiguousErrors(false)`

`ambiguousErrors` means that ambiguity errors *must* be reported,
even when in silent mode. When it's false, they are *not* reported,
but they are buffered when the context reporter is buffering.

TODO: this seems a bit dubious, but this is what happens now.
Let's see if we can simplify this once the refactoring is complete.

Again, the end goal is a strategy-based approach to reporting,
where the reporting mode is captured in the reporter being used,
with as little mutation as possible to capture more invariants
(would like to stop throwing TypeError eventually and only have two reporters:
buffering reporter, regular reporter)